+++ src/main/java/com/shaadi/dto/PurchaseSubscriptionDto.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class PurchaseSubscriptionDto {
-    private Integer planId;
+    private Long planId;
 }
